tests/wpt/tests/css/css-break/out-of-flow-in-multicolumn-108.html

<!DOCTYPE html>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://bugs.chromium.org/p/chromium/issues/detail?id=1399313">
<meta name="assert" content="For a non-OOF variant of this test, see //css/css-break/overflow-clip-007.html">
<style>
  body { margin: 0; }
</style>
<div style="columns:3; width:100px; gap:20px; column-fill:auto; height:100px;">
  <div style="overflow:clip;">
    <div style="position:relative; height:300px;">
      <div style="position:absolute; width:100%; height:100000px;">
        <div style="height:10050px;"></div>
        <div id="target" style="height:1000px;"></div>
      </div>
    </div>
  </div>
</div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(()=> {
    var rect = target.getBoundingClientRect();
    assert_equals(rect.left, 80);
    assert_equals(rect.top, 9850);
    assert_equals(rect.width, 20);
    assert_equals(rect.height, 1000);
  }, "Overflowing and clipped content doesn't fragment");
</script>
